Extract some parts of Program to RawProgram
